01-rag-chatbot/rag-flow/requirements.txt
promptflow>=1.10.0
promptflow-tools>=1.4.0
azure-identity>=1.15.0
azure-search-documents>=11.4.0
openai>=1.12.0
